tools/contractor-review-agent/src/dspy/requirements.txt
dspy-ai>=2.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
